experiments/test_order_flow.py
"""
Test Order Flow Experiment

This script explores order placement mechanics without executing actual trades.
It validates order structures, checks minimums, and simulates the order lifecycle.
"""

import asyncio
import os

import ccxt.async_support as ccxt
from dotenv import load_dotenv

load_dotenv()

API_KEY = os.getenv("BINANCE_API_KEY")
API_SECRET = os.getenv("BINANCE_API_SECRET")


async def analyze_order_structure():
    """Analyze what an order structure looks like and minimum requirements."""
    print("=" * 60)
    print("ORDER FLOW ANALYSIS")
    print("=" * 60)

    exchange = ccxt.binance({
        'apiKey': API_KEY,
        'secret': API_SECRET,
        'enableRateLimit': True,
    })

    try:
        await exchange.load_markets()

        # Analyze BTC/USDT market
        symbol = 'BTC/USDT'
        market = exchange.markets[symbol]

        print(f"\nMarket: {symbol}")
        print(f"  Min order amount: {market['limits']['amount']['min']} BTC")
        print(f"  Max order amount: {market['limits']['amount']['max']} BTC")
        print(f"  Min cost: {market['limits']['cost']['min']} USDT")
        print(f"  Price precision: {market['precision']['price']} decimals")
        print(f"  Amount precision: {market['precision']['amount']} decimals")

        # Check our balance
        balance = await exchange.fetch_balance()
        btc_balance = balance['BTC']['free']
        usdc_balance = balance['USDC']['free']

        print("\nOur balances:")
        print(f"  BTC: {btc_balance}")
        print(f"  USDC: {usdc_balance}")

        # Calculate min order in USDC terms
        ticker = await exchange.fetch_ticker(symbol)
        current_price = ticker['last']
        min_btc = market['limits']['amount']['min']
        min_cost_usd = min_btc * current_price

        print(f"\nOrder minimums at price {current_price}:")
        print(f"  Min BTC amount: {min_btc}")
        print(f"  Equivalent USD: ~{min_cost_usd:.2f}")
        print(f"  Min order cost: {market['limits']['cost']['min']} USDT")

        # Check if we can place a minimum order
        print("\nCan we place a minimum order?")
        if usdc_balance >= min_cost_usd:
            print("  ✓ Yes - sufficient USDC balance")
        else:
            print(f"  ✗ No - need {min_cost_usd:.2f} USDC, have {usdc_balance:.2f}")

        # Explore create_order parameters without executing
        print("\n" + "=" * 60)
        print("ORDER CREATION PARAMETERS")
        print("=" * 60)

        print(f"""
To place a market buy order for {symbol}:

Method: exchange.create_market_buy_order(symbol, amount)
  - symbol: '{symbol}' (CCXT format)
  - amount: BTC amount to buy (e.g., {min_btc})

OR using create_order for more control:

Method: exchange.create_order(
    symbol='{symbol}',
    type='market',
    side='buy',
    amount={min_btc},
    params={{}}
)

Expected response structure:
{{
    'id': 'order_id_string',
    'clientOrderId': 'client_assigned_id',
    'timestamp': 1234567890000,
    'datetime': '2026-01-29T15:00:00.000Z',
    'symbol': '{symbol}',
    'type': 'market',
    'side': 'buy',
    'price': None,           # Market order - no fixed price
    'amount': {min_btc},     # Requested amount
    'cost': 100.00,          # Actual USDT spent
    'filled': {min_btc},     # Amount filled
    'remaining': 0.0,
    'status': 'closed',      # Filled immediately
    'trades': [...]          # List of fills
}}
        """)

        # Get exchange info for trading rules
        print("\n" + "=" * 60)
        print("EXCHANGE TRADING RULES")
        print("=" * 60)

        # Check if we can use the test order endpoint (validation only)
        print("\nBinance test order endpoint:")
        print("  - Use params={'test': True} to validate without execution")
        print("  - This checks: balance, min amounts, filters")
        print("  - Does NOT actually place the order")

        return {
            'min_btc': min_btc,
            'min_cost_usd': min_cost_usd,
            'current_price': current_price,
            'can_trade': usdc_balance >= min_cost_usd
        }

    finally:
        await exchange.close()


async def simulate_order_lifecycle():
    """Simulate the full order lifecycle for documentation."""
    print("\n" + "=" * 60)
    print("ORDER LIFECYCLE SIMULATION")
    print("=" * 60)

    print("""
Step 1: Strategy generates signal
  - Strategy.run() returns target_position (0.0 to 1.0)
  - Executor calculates: target_value = total_value * target_position
  - Executor calculates: current_value = base_amount * current_price
  - Executor calculates: delta = target_value - current_value

Step 2: Executor determines order parameters
  if delta > 0:
    - Action: BUY base currency
    - Side: 'buy'
    - Amount: delta / current_price
  elif delta < 0:
    - Action: SELL base currency
    - Side: 'sell'
    - Amount: abs(delta) / current_price
  else:
    - No action needed

Step 3: Validate order against market limits
  - Check amount >= market.limits.amount.min
  - Check amount <= market.limits.amount.max
  - Check cost >= market.limits.cost.min
  - Round amount to market.precision.amount

Step 4: Check balance sufficiency
  - BUY: Check quote_balance >= estimated_cost
  - SELL: Check base_balance >= amount

Step 5: Submit order
  - Use create_market_buy_order() or create_market_sell_order()
  - Set clientOrderId for idempotency (optional)
  - Handle rate limits and retries

Step 6: Handle response
  if status == 'closed':
    - Order filled immediately
    - Update strategy state with actual fill amounts
    - Log trade details
  elif status == 'open':
    - For market orders, this shouldn't happen
    - Poll order status until filled
  elif status == 'rejected' or status == 'canceled':
    - Log error
    - Strategy state unchanged

Step 7: Update database
  - Insert trade record
  - Update strategy position
  - Update wallet snapshot
""")


async def main():
    """Run order flow experiments."""
    print("\n" + "=" * 60)
    print("ORDER FLOW EXPERIMENT")
    print("=" * 60)
    print("\n⚠️  This experiment does NOT place actual orders")
    print("   It only analyzes parameters and validates requirements.\n")

    if not API_KEY or not API_SECRET:
        print("✗ ERROR: API keys not found!")
        return

    result = await analyze_order_structure()
    await simulate_order_lifecycle()

    print("\n" + "=" * 60)
    print("EXPERIMENT COMPLETE")
    print("=" * 60)
    print(f"\nCan place minimum order: {'YES' if result['can_trade'] else 'NO'}")
    if not result['can_trade']:
        print(f"Need ~${result['min_cost_usd']:.2f} USDC for minimum BTC order")


if __name__ == "__main__":
    asyncio.run(main())
